upload, delete the cast

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -10,8 +10,10 @@ hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *hash_t = NULL;
 	unsigned long int i = 0;
-	if ((int)size <= 0)
+
+	if (size <= 0)
 		return (NULL);
+
 	hash_t =  malloc(sizeof(hash_table_t));
 	if (hash_t == NULL)
 		return (NULL);
